[feat/LIG-58] 1. move proto unittest from common -> proto 2. modify workflow: should test proto package
