+++ src/proto/Cargo.toml
@@ -24,3 +24,6 @@ all = ["worker", "coordinator", "proto-common"]
 
 [build-dependencies]
 tonic-build = "0.8"
+
+[dev-dependencies]
+serde_json = "1.0.59"
